+++ modules/ip_info/ip_info.py
@@ -480,7 +480,8 @@ class IPInfo(IModule, multiprocessing.Process):
         wait for update manager to finish updating the mac db and open the rest of dbs before starting this module
         """
         # this is the loop that controls te running on open_dbs
-        loop = asyncio.get_event_loop()
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
         # run open_dbs in the background so we don't have
         # to wait for update manager to finish updating the mac db to start this module
         loop.run_until_complete(self.open_dbs())